DICOM gadget made 32/64 bit safe, and DICOM writing ability added to mriclient.
Former-commit-id: 2625c11eed1c9774a597257aed93822f566aa9bf

+++ gadgets/dicom/DicomImageWriter.cpp
@@ -81,7 +81,7 @@ END DEBUG */
     //GADGET_DEBUG2("Sent GadgetMessageIdentifier %d\n", id.id);
 
 
-    unsigned long nbytes = (unsigned long)serialized_length;
+    uint32_t nbytes = (uint32_t)serialized_length;
     //GADGET_DEBUG2("Sending bytes length %d\n", serialized_length);
     if ((send_cnt = sock->send_n (&nbytes, sizeof(nbytes))) <= 0) {
         ACE_DEBUG ((LM_ERROR, ACE_TEXT ("(%P|%t) Unable to send DICOM bytes length\n")));
